requirements.txt
ipykernel
python-dotenv
pydantic_ai
langgraph